# Conflicts:
#	CHANGELOG-7.5.md
#	source/Application/Controller/Admin/LoginController.php
#	source/Application/Model/Order.php
#	source/Core/Database/Adapter/Doctrine/Database.php
#	source/Core/Exception/DatabaseNotConfiguredException.php
#	source/Core/Exception/StandardException.php
#	source/Core/Model/ListModel.php
#	source/Core/UniversallyUniqueIdGenerator.php
#	source/Core/UtilsCount.php
#	source/Internal/Framework/Console/Executor.php
#	source/Internal/Framework/Console/ExecutorInterface.php
#	source/Internal/Framework/Database/Logger/QueryLogger.php
#	source/Internal/Framework/FileSystem/FileGenerator/CsvFileGenerator.php
#	source/Internal/Setup/Database/Service/DatabaseCreator.php
#	source/Internal/Setup/Database/Service/DatabaseInitiator.php
#	source/Setup/Database.php
#	source/Setup/Exception/CommandExecutionFailedException.php
#	source/Setup/Exception/LanguageParamsException.php
#	source/Setup/Exception/SetupControllerExitException.php
#	source/Setup/Exception/TemplateNotFoundException.php
#	source/Setup/Utilities.php
#	tests/CachingTrait.php
#	tests/Integration/Internal/Domain/Review/Bridge/ProductRatingBridgeTest.php
#	tests/Integration/Internal/Domain/Review/Dao/ProductRatingDaoTest.php
#	tests/Integration/Internal/Framework/Database/QueryBuilderFactoryTest.php
#	tests/Integration/Internal/Setup/Database/DatabaseCreatorTest.php
#	tests/Integration/Legacy/Core/Database/Adapter/DatabaseInterfaceImplementation.php

+++ source/Application/Component/Widget/ArticleDetails.php
@@ -703,14 +703,14 @@ class ArticleDetails extends \OxidEsales\Eshop\Application\Component\Widget\Widg
     public function getRatingValue()
     {
         if ($this->_dRatingValue === null) {
-            $this->_dRatingValue = (double) 0;
+            $this->_dRatingValue = 0.0;
             if ($this->isReviewActive() && ($oDetailsProduct = $this->getProduct())) {
                 $blShowVariantsReviews = \OxidEsales\Eshop\Core\Registry::getConfig()->getConfigParam('blShowVariantReviews');
                 $this->_dRatingValue = round($oDetailsProduct->getArticleRatingAverage($blShowVariantsReviews), 1);
             }
         }
 
-        return (double) $this->_dRatingValue;
+        return (float) $this->_dRatingValue;
     }
 
     /**
